Order Management Implemented

+++ backend/src/main/kotlin/com/lvlup/backend/dto/PaginatedDataResponse.kt
@@ -2,8 +2,8 @@ package com.lvlup.backend.dto
 
 data class PaginatedDataResponse<T>(
     val data: List<T>,
-    val totalElements: Int,
-    val totalPages: Int,
+    val totalElements: Long,
+    val totalPages: Long,
     val currentPage: Int,
     val pageSize: Int
 )
